+++ get/install.bash
@@ -6,7 +6,7 @@
 #   $ curl -sL https://get.webexec.sh -o get-webexec.sh
 #   $ ./get-webexec.sh
 #
-LATEST_VERSION="0.17.3"
+LATEST_VERSION="0.17.4"
 
 # The latest release is currently hard-coded.
 echo ">>> Installing webexec latest version"