+++ CHANGELOG.md
@@ -8,7 +8,7 @@ his file's format is define in
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and the release workflow reads it to set github's release notes.
 
-## Unreleased
+## [0.17.4] 2022-08-13
 
 ### Fixed